LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/the-offscreen-canvas/size.attributes.parse.onlyspace-expected.txt

size.attributes.parse.onlyspace

Parsing of non-negative integers


FAIL Parsing of non-negative integers assert_equals: offscreenCanvas.width === 100 (got 0[number], expected 100[number]) expected 100 but got 0
